improvements

+++ make_collage_shader.py
@@ -6,8 +6,8 @@ code_path = str(Path(workspace(q=True, rd=True)) / 'shading' / 'fantasy' / 'code
 if code_path not in sys.path:
     sys.path.append(code_path)
 
-import subprocess, webbrowser
-from internals import network, shading_path, surface_values, coordinate_converter, palettes, collage_shader, world_placement, screen_placement, tracking_projection
+import subprocess
+from internals import network, shading_path, surface_values, coordinate_converter, palettes, collage_shader, world_placement, screen_placement, tracking_projection, dialog_with_support
 importlib.reload(network)
 importlib.reload(shading_path)
 importlib.reload(collage_shader)
@@ -17,11 +17,15 @@ importlib.reload(surface_values)
 importlib.reload(palettes)
 importlib.reload(world_placement)
 importlib.reload(screen_placement)
+importlib.reload(dialog_with_support)
+from internals.dialog_with_support import dialog_with_support
 from internals.shading_path import shading_path
 from internals.surface_values import calculate_surface_values
 from internals.collage_shader import CollageShader
 from pathlib import Path
+from hashlib import md5
 import json
+from shutil import rmtree
 
 
 selection = ls(sl=True)
@@ -34,31 +38,26 @@ if dialog_output is None:
     exit()
 map_image_path = Path(dialog_output[0])
 
+image_hash = md5(map_image_path.read_bytes()).hexdigest()
+
 map_dir_path = map_image_path.with_name(map_image_path.stem)
 map_data_path = map_dir_path / 'map data.json'
 masks_path = map_dir_path / 'masks'
 
-new_map_data = not map_data_path.exists()
+if map_data_path.exists():
+    with map_data_path.open() as file:
+        original_map_data = json.load(file)
+    if original_map_data['hash'] == image_hash:
+        map_data_status = 'up to date'
+    else:
+        map_data_status = 'out of date'
+else:
+    map_data_status = 'nonexistent'
 
 if not map_dir_path.exists():
     map_dir_path.mkdir()
 
-def dialog_with_support(title, msg, buttons, **kwargs):
-    dialog_output = confirmDialog(t=title, m=msg, b=['What?'] + buttons, ma='left', **kwargs)
-    if dialog_output == 'What?':
-        webbrowser.open('slack://channel?id=C05BFV55GLT&team=T05B9C5MHKQ')
-        dialog_output = confirmDialog(t='Would you like to copy the error message to your clipboard?', m=title + '\n' + msg, b=['Yes', 'No'], cb='No', db='Yes', icon='question', ma='left')
-        if dialog_output == 'Yes':
-            if sys.platform == 'darwin':
-                copy_keyword = 'pbcopy'
-            elif sys.platform == 'win32':
-                copy_keyword = 'clip'
-            subprocess.run(copy_keyword, universal_newlines=True, input=title + '\n' + msg)
-        exit()
-    else:
-        return dialog_output
-
-if new_map_data:
+if map_data_status != 'up to date':
     map_proc = subprocess.run(['python3', shading_path('code', 'internals', 'map_data.py'), map_image_path, map_data_path], capture_output=True)
     if map_proc.returncode or (err := map_proc.stderr.decode("utf-8")) != '':
         err = map_proc.stderr.decode("utf-8")
@@ -68,11 +67,14 @@ if new_map_data:
             title = 'Error'
             msg = err
         dialog_output = dialog_with_support(title, msg, ['I’ll fix it'], cb='I’ll fix it', db='I’ll fix it', icon='warning')
-        map_data_path.unlink(missing_ok=True)
+        if map_data_status == 'nonexistent':
+            map_data_path.unlink(missing_ok=True)
         exit()
 
 with map_data_path.open() as file:
     map_data = json.load(file)
+
+map_data['hash'] = image_hash
 
 if map_data['anti-aliasing warning']:
     msg = 'It looks like you may have used a brush with anti-aliasing, which produces many colors along the edge of a stroke. If that is the case and you make a shader based on this map, each of those different colors along the edge will become its own facet. Do you still want to continue?'
@@ -80,30 +82,37 @@ if map_data['anti-aliasing warning']:
     if dialog_output in ['No', 'dismiss']:
         map_data_path.unlink()
         exit()
+    else:
+        map_data['anti-aliasing warning'] = False
+
+facet_borders_changed = map_data_status != 'nonexistent' and original_map_data['pixels'] != map_data['pixels']
+blur_markers_changed = map_data_status != 'nonexistent' and [facet['blur markers'] for facet in original_map_data['facets'].values()] != [facet['blur markers'] for facet in map_data['facets'].values()]
 
 num_facets = len(map_data['facets'])
-if num_facets > 1:
-    if new_map_data:
-        promptDialog(t='Enter Resolution', m='Enter the resolution for the blur mask images. Larger values take longer to run.', b=['OK'], ma='left', st='integer', tx='256')
-        blur_resolution = promptDialog(q=True, tx=True)
-        map_data['blur resolution'] = int(blur_resolution)
-        with map_data_path.open('w') as file:
-            json.dump(map_data, file, indent=4)
-
-    if not masks_path.exists():
-        node = selection[0]
+if num_facets > 1 and (not masks_path.exists() or facet_borders_changed or blur_markers_changed):
+    if masks_path.exists():
+        rmtree(masks_path)
+    promptDialog(t='Enter Resolution', m='Enter the resolution for the blur mask images. Larger values take longer to run.', b=['OK'], ma='left', st='integer', tx='256')
+    blur_resolution = int(promptDialog(q=True, tx=True))
+    node = None
+    for node in selection:
         if node.type() == 'mesh':
             obj = node.getTransform()
+            break
         elif node.type() == 'transform' and node.getShape().type() == 'mesh':
             obj = node
-        surface_values_path = calculate_surface_values(obj, map_data_path)
-        blur_proc = subprocess.run(['python3', shading_path('code', 'internals', 'blur_images.py'), surface_values_path, map_data_path], capture_output=True)
-        print('*' * 30, blur_proc.stdout, blur_proc.stderr)
-else:
-    if 'blur resolution' in map_data:
-        del map_data['blur resolution']
-        with map_data_path.open('w') as file:
-            json.dump(map_data, file, indent=4)
+            break
+    if node is None:
+        dialog_with_support('Error', 'Please select at least one mesh.', ['OK'], cb='OK', db='OK', icon='warning')
+    if node.type() == 'mesh':
+        obj = node.getTransform()
+    elif node.type() == 'transform' and node.getShape().type() == 'mesh':
+        obj = node
+    surface_values_path = calculate_surface_values(obj, map_data_path, blur_resolution)
+    blur_proc = subprocess.run(['python3', shading_path('code', 'internals', 'blur_images.py'), surface_values_path, map_data_path], capture_output=True)
+
+with map_data_path.open('w') as file:
+    json.dump(map_data, file, indent=4)
 
 for node in selection:
     if node.type() == 'mesh':
